+++ tic-tac-toe-challenge/src/App.css
@@ -63,6 +63,7 @@ header {
   display: flex;
   justify-content: space-between;
   margin: 30px 0;
+  flex-direction: column;
 }
 
 .player {
