+++ tic-tac-toe-challenge/src/utils.ts
@@ -3,3 +3,19 @@ export const formatTime = (time: number) => {
     const seconds = time % 60;
     return `${minutes}:${seconds < 10 ? '0' + seconds : seconds}`;
 }
+
+export const checkWinner = (board: (string | null)[]): [string | null, number[] | null] => {
+    const winningCombinations = [
+      [0, 1, 2], [3, 4, 5], [6, 7, 8],
+      [0, 3, 6], [1, 4, 7], [2, 5, 8],
+      [0, 4, 8], [2, 4, 6]
+    ];
+  
+    for (let combination of winningCombinations) {
+      const [a, b, c] = combination;
+      if (board[a] && board[a] === board[b] && board[a] === board[c]) {
+        return [board[a], combination]; // Return the winner and the winning combination
+      }
+    }
+    return [null, null];
+  };
